move methods into a sessions helper
